requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
requests==2.32.3
pandas==2.2.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pdfplumber==0.11.0
python-docx==1.1.0
nltk==3.9.1
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.7.4
typing_extensions==4.12.2
python-dotenv==1.0.1
